complete1

+++ app/Http/Controllers/TasksController.php
@@ -11,11 +11,21 @@ class TasksController extends Controller
     // getでTask/にアクセスされた場合の「一覧表示処理」
     public function index()
     {
+         // タスク一覧を取得
+            $tasks = Task::all();
         // ここで認証状態によって分岐する コントローラーにかく
         if (\Auth::check()) {
             // 認証済みの場合はここ
-            // タスク一覧を取得
-            $tasks = Task::all();
+             $user = \Auth::user();
+             $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+             $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+          
+            // ログインユーザーのタスクのみ全権取得する必要がある
+            $tasks = \Auth::user()->tasks;
+           
              // タスク一覧ビューでそれを表示
             return view('tasks.index', [
                 'tasks' => $tasks,
@@ -32,7 +42,8 @@ class TasksController extends Controller
         $task = new Task;
 
         // タスク作成ビューを表示
-        return view('tasks.create', [
+        
+            return view('tasks.create', [
             'task' => $task,
         ]);
     }
@@ -40,18 +51,23 @@ class TasksController extends Controller
     // postでmessages/にアクセスされた場合の「新規登録処理」
     public function store(Request $request)
     {
-        
         // バリーデーション
         $request->validate([
             'status' => 'required|max:10',   // 追加
             'content' => 'required|max:255',
         ]);
-        
+    
         // タスクを作成
+       
         $tasks = new Task;
         $tasks->status = $request->status;
         $tasks->content = $request->content;
+        $tasks->user_id = \Auth::user()->id;
+
         $tasks->save();
+        
+
+        
 
         // トップページへリダイレクトさせる
         return redirect('/');
@@ -63,11 +79,16 @@ class TasksController extends Controller
     {
         // idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
-
-        // メッセージ詳細ビューでそれを表示
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+        
+        if ($task->user_id === \Auth::user()->id) {
+            // メッセージ詳細ビューでそれを表示
+            return view('tasks.show', [
+                'task' => $task,
+            ]);
+        } else {
+            // ログインユーザーではないユーザーのタスクなので、詳細ページは表示しない
+               return redirect('/');
+        }
     }
 
    // getでmessages/id/editにアクセスされた場合の「更新画面表示処理」
@@ -77,9 +98,14 @@ class TasksController extends Controller
         $task = Task::findOrFail($id);
 
         // メッセージ編集ビューでそれを表示
-        return view('tasks.edit', [
+        if ($task->user_id === \Auth::user()->id) {
+            return view('tasks.edit', [
             'task' => $task,
         ]);
+        } else {
+            // ログインユーザーではないユーザーのタスクなので、詳細ページは表示しない
+               return redirect('/');
+     }
     }
 
     // putまたはpatchでmessages/idにアクセスされた場合の「更新処理」
@@ -94,10 +120,11 @@ class TasksController extends Controller
         $task = Task::findOrFail($id);
         
         // メッセージを更新
+       if (\Auth::check()){
         $task->status = $request->status;    // 追加
         $task->content = $request->content;
         $task->save();
-
+       }
         // トップページへリダイレクトさせる
         return redirect('/');
     }
@@ -108,9 +135,11 @@ class TasksController extends Controller
         // idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
         // メッセージを削除
-        $task->delete();
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
 
         // トップページへリダイレクトさせる
-        return redirect('/');
+       return redirect('/');
     }
 }
